+++ src/Libraries/SmartStore.Data/Mapping/Payments/PaymentMethodMap.cs
@@ -16,7 +16,11 @@ namespace SmartStore.Data.Mapping.Payments
 			this.Property(x => x.ExcludedCountryIds).IsMaxLength();
 			this.Property(x => x.ExcludedShippingMethodIds).IsMaxLength();
 
+			this.Property(x => x.MinimumOrderAmount).HasPrecision(18, 4);
+			this.Property(x => x.MaximumOrderAmount).HasPrecision(18, 4);
+
 			this.Ignore(x => x.CountryExclusionContext);
+			this.Ignore(x => x.AmountRestrictionContext);
 		}
 	}
 }