+++ src/Libraries/SmartStore.Data/Migrations/201506181501470_PaymentShippingRestrictions.Designer.cs
@@ -13,7 +13,7 @@ namespace SmartStore.Data.Migrations
         
         string IMigrationMetadata.Id
         {
-            get { return "201506170941137_PaymentShippingRestrictions"; }
+            get { return "201506181501470_PaymentShippingRestrictions"; }
         }
         
         string IMigrationMetadata.Source