+++ src/Libraries/SmartStore.Data/Migrations/201506181501470_PaymentShippingRestrictions.cs
@@ -18,6 +18,9 @@ namespace SmartStore.Data.Migrations
                         ExcludedCountryIds = c.String(),
                         ExcludedShippingMethodIds = c.String(),
                         CountryExclusionContextId = c.Int(nullable: false),
+                        MinimumOrderAmount = c.Decimal(precision: 18, scale: 4),
+                        MaximumOrderAmount = c.Decimal(precision: 18, scale: 4),
+                        AmountRestrictionContextId = c.Int(nullable: false),
                     })
                 .PrimaryKey(t => t.Id)
 				.Index(t => t.PaymentMethodSystemName);
@@ -47,13 +50,13 @@ namespace SmartStore.Data.Migrations
 				"Restrictions",
 				"Einschränkungen");
 
-			builder.AddOrUpdate("Admin.Common.AppliedTo",
-				"applied to",
+			builder.AddOrUpdate("Admin.Common.RelatedTo",
+				"related to",
 				"bezogen auf");
 
 			builder.AddOrUpdate("Admin.Configuration.Payment.Methods.RestrictionNote",
 				"Select customer roles, shipping methods and countries for which you do <u>not</u> want to offer this payment method.",
-				"Wählen Sie Kundengruppen, Versandarten und Länder, bei denen sie diese Zahlungsmethode <u>nicht</u> anbieten möchten.");
+				"Wählen Sie Kundengruppen, Versandarten und Länder, bei denen Sie diese Zahlungsmethode <u>nicht</u> anbieten möchten.");
 
 			builder.AddOrUpdate("Enums.SmartStore.Core.Domain.Payments.CountryExclusionContextType.BillingAddress",
 				"Billing address",
@@ -65,7 +68,33 @@ namespace SmartStore.Data.Migrations
 
 			builder.AddOrUpdate("Admin.Configuration.Shipping.Methods.RestrictionNote",
 				"Select customer roles for which you do <u>not</u> want to offer this shipping method.",
-				"Wählen Sie Kundengruppen, bei denen sie diese Versandart <u>nicht</u> anbieten möchten.");
+				"Wählen Sie Kundengruppen, bei denen Sie diese Versandart <u>nicht</u> anbieten möchten.");
+
+			builder.AddOrUpdate("Admin.Configuration.Payment.Methods.MinimumOrderAmount",
+				"Minimum order amount",
+				"Mindestbestellwert",
+				"Specifies the minimum order amount from which on the payment method should be offered.",
+				"Legt den Mindestbestellwert fest, ab dem die Zahlungsmethode angeboten werden soll.");
+
+			builder.AddOrUpdate("Admin.Configuration.Payment.Methods.MaximumOrderAmount",
+				"Maximum amount",
+				"Maximalbestellwert",
+				"Specifies the maximum order amount up to which the payment methods should be offered.",
+				"Legt den maximalen Bestellwert fest, bis zu dem die Zahlungsmethode angeboten werden soll.");
+
+			builder.AddOrUpdate("Admin.Configuration.Payment.Methods.AmountRestrictionContext",
+				"Amount related to",
+				"Betrag bezieht sich auf",
+				"Specifies the amount to which the minimum and maximum order amounts are related to.",
+				"Legt den Betrag fest, auf den sich die Minimal- und Maximalbestellwerte beziehen.");
+
+			builder.AddOrUpdate("Enums.SmartStore.Core.Domain.Payments.AmountRestrictionContextType.SubtotalAmount",
+				"Subtotal",
+				"Zwischensumme");
+
+			builder.AddOrUpdate("Enums.SmartStore.Core.Domain.Payments.AmountRestrictionContextType.TotalAmount",
+				"Order total",
+				"Gesamtsumme");
 		}
     }
 }